refactor(experience): update MFA error handling
